ECS implementation -
an entity is just an id, which sole purpose is to group components logically together.

a component is just a store for data (a ‘struct’ without code logic).

a system is responsible for processing / transforming the data stored in the components; to achieve this, it iterates over homogeneous groups of components.